requirements.txt
laserforce.py==1.0.4
openskill==5.1.0
bcrypt==3.2.0
sanic-cors==2.2.0
sanic==25.3.0
sanic-ext==24.12.0
tortoise-orm==0.19.2
sanic_session==0.8.0
Jinja2==3.1.2
sentry_sdk==1.25.0
sanic-jinja2==2022.1.13
aioredis==2.0.1
aerich==0.7.1
pytest==8.4.1
pytest-cov==6.2.1
flake8==7.0.0
websockets==10.0
pytailwind==0.0.3
readline